refactor: migrate VersionService and related configs to client app
- Removed shared library VersionService and tests.
- Added VersionService and updated tests in the client app, integrating with environment configuration.
- Adjusted Jest setup and removed unused configuration from the shared library.
