+++ apps/client/src/app/app.config.ts
@@ -14,8 +14,6 @@ import {
     withFetch,
     withInterceptorsFromDi,
 } from '@angular/common/http';
-import { VERSION_CONFIG } from '@drevo-web/shared';
-import { environment } from '../environments/environment';
 
 const routesTracing = false;
 
@@ -28,9 +26,5 @@ export const appConfig: ApplicationConfig = {
         importProvidersFrom(
             RouterModule.forRoot(appRoutes, { enableTracing: routesTracing })
         ),
-        {
-            provide: VERSION_CONFIG,
-            useValue: { version: environment.version }
-        },
     ],
 };
