+++ apps/client/src/app/services/version/version.service.ts
@@ -1,0 +1,9 @@
+import { Injectable } from '@angular/core';
+import { environment } from '../../../environments/environment';
+
+@Injectable({ providedIn: 'root' })
+export class VersionService {
+    getVersion(): string {
+        return environment.version;
+    }
+}
