+++ libs/shared/package.json
@@ -6,7 +6,6 @@
     "main": "./src/index.js",
     "types": "./src/index.d.ts",
     "dependencies": {
-        "tslib": "^2.3.0",
-    "@angular/core": "~19.2.0"
+        "tslib": "^2.3.0"
     }
 }
